agent/agent_requirements.txt
requests==2.31.0
Pillow==10.1.0
schedule==1.2.0